Fix the test mkey_test_sig_incorrect_ref_tag
RDMA_WRITE is a used data tranfer operation in the test. The data path
is showed below:

src mkey domain -> src wire domain -> dst wire domain -> dst mkey domain

The t10dif is validated at two points: src mkey domain and dst wire domain.

The test corrupts reftag in the src wire domain and expects a signature
error in the src mkey. But the signature is not validated there and
the error cannot happen.

Corrupt reftag in the mkey domain and not in the wire domain.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>