Remove the non_inline tests
The non_inline feature is not supported, and we are not going to
support it now. The related tests are disabled in the sources. That's
why the test report always has a notification about disabled tests. It
is confusing because it is not clear which tests are disabled, and the
same notification is shown if the HCA capabilities are not correct.

Remove the non_inline tests. We have the git history, and we can easy
restore the tests further.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>