Fix failed tests layouts/mkey_test_basic/*.non_signaled
The test fixture has two sides: src and dst. Each side has its own QP
and MKEY. The test configures the MKEYs. Each MKEY is configured without
the send signal flag via QP from the appropriate side. So, the MKEYs are
configured on the different QPs. Then the test performs an RDMA
operation. If the operation type is "read", it is posted to the dst QP;
otherwise, it is posted to the src QP.

The test can fail because one of the MKEYs is not always ready at the
RDMA operation. That happens because WQEs in the dst and src QPs are
concurrent, and it is not defined which one is finished first. It is
allowed to drop the signal flag and do not wait for completion only if
MKEY configure and RDMA operation are posted to the same QP. And we have
to wait for completion for MKEY configure from the other QP.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>